twitter_monitoring/requirements.txt
tweepy==3.10.0
python-dotenv
boto3